+++ CHANGELOG.md
@@ -7,6 +7,9 @@ All notable changes to this project will be documented in this file.
 
 ## [2.0.10] - 2025-04-2
 
+### Added
+- Added keyboard shortcuts ('Y' to accept, 'N' to decline) for responding to lobby invitations
+
 ### Fixed
 - Fixed a bug where match features did not apply after the first match when following a board